Move creation button to the topbar (#116)
* Refactor org creation modal to its own file

* Move creation button to the topbar

* Add CHANGELOG entry

* Refactor Custom Field duplicated code
