+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Move the organization creation button to the topbar
+
 ## [1.24.7] - 2023-07-13
 
 ### Changed
